Switch from --devel to --HEAD (#13)
Instead of offering to build the latest nightly, we now offer to build the latest master. Also, Travis for this tap was broken and this commit fixes it. Note that `brew upgrade python` and `brew upgrade automake` are a workaround for failing --HEAD builds on Travis where the installation would succeed but return 1.